Change the default configuration for Fetchers:
  tez.runtime.shuffle.connect.timeout = 27500

The following configuration keys use their default values:

  tez.runtime.shuffle.speculative.fetch.wait.millis (12500)
  tez.runtime.shuffle.stuck.fetcher.threshold.millis (2500)
  tez.runtime.shuffle.stuck.fetcher.release.millis (7500)

If a remote ContainerWorker is unreachable, InputReadErrorEvent is sent in 30 seconds
(because tez.runtime.shuffle.connect.timeout = 27500).

If a remote ContainerWorker is reachable but unstable, up to 3 Fetcher can be created
within 30 seconds before InputReadErrorEvent is reported to DAGAppMaster.
  - worst case: 0ms, 12500ms, 25000ms
  - best case: 0ms, 7500ms, 15000ms

In an environment where ContainerWorkers are killed frequently, setting
tez.runtime.shuffle.connect.timeout to 2500 can be a reasonable choice.
If it is set to 2500, a connection failure immediately reports InputReadErrorEvent
and new Tasks are executed in the upstream Vertex (with Vertex rerun).